Refactor(ble): ble 폴더 SPI 송신 경로 uint8_t 전환 (B3a)
송신 경로는 호출부가 57곳이라 한 번에 못 옮긴다. 파일 단위로 이행할 수
있도록 신규 API 를 병행 도입하고 ble 폴더부터 옮겼다.

HAL
- tdc_hal_spi_write_tx_buffer_u8(const uint8_t*, int) 신규 추가
- 기존 tdc_hal_spi_write_tx_buffer(int*, int) 는 전환 기간 어댑터로 남긴다.
  내부에서 원소별 변환 복사 후 신규 API 를 호출한다. 캐스팅이 아니라
  변환 복사여야 한다 - uint8_t 버퍼를 int* 로 넘기면 호출자가 4바이트씩
  읽어 즉시 깨진다. 전 호출부 이행 후 B3c 에서 제거한다.

ble 폴더 (8파일)
- 로컬 조립 버퍼 6개를 uint8_t 로: tdc_ble_communication.c(Tx_dataBuff),
  tdc_ble_mapping.c 2곳, tdc_ble_remote.c 2곳, tdc_ble_remote_sp_para.c
- 송신 호출 33곳을 _u8 로 교체
- tdc_ble_general_debug.c/.h, tdc_ble_gain_control.c/.h 의 tx_buf 인자를
  uint8_t* 로 (tdc_ble_remote.c 가 자기 버퍼를 넘기는 구조라 연쇄)

효과
- text 185772 -> 183920 (1852 바이트 감소). uint8_t 접근이 더 짧은
  명령으로 컴파일되고 스택 사용도 줄었다.
- ble 폴더에 int 조립 버퍼와 구 API 호출이 0건 남았다.

검증
- CLI 빌드 에러 0 / 경고 0
- 실기 확인 대기 (isd/dfu/sys 는 아직 구 API 경유라 동작 동일해야 한다)

+++ src/2__cm3/source/ble/tdc_ble_general_debug.c
@@ -25,14 +25,14 @@
 #include <tdc_dfu_ble_ota.h>
 #include <tdc_fs_map.h>
 
-static int gd_handle_touch_debug(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option);      // option 1
-static int gd_handle_no_backtel(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option);       // option 2
-static int gd_handle_map_init(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option);         // option 3
+static int gd_handle_touch_debug(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option);      // option 1
+static int gd_handle_no_backtel(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option);       // option 2
+static int gd_handle_map_init(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option);         // option 3
 /* option 4(운영 모드 설정)는 fake sleep mode 제거와 함께 삭제됨(2026-07-20).
  * 미지정 option 은 아래 else 에서 command 만 loop-back 하고 무시된다.
  * 상세: docs/tasks/main/20260720_fake-sleep-removal/ */
 
-int tdc_ble_general_debug_handle(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index)
+int tdc_ble_general_debug_handle(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index)
 {
     int option = packet->data[0];  // 옵션
 
@@ -59,7 +59,7 @@ int tdc_ble_general_debug_handle(const ST__REMOTECONTROL_PACKET *packet, int *tx
     return tx_index;
 }
 
-static int gd_handle_touch_debug(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option)
+static int gd_handle_touch_debug(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option)
 {
     uint16_t lta;
     uint16_t count;
@@ -99,7 +99,7 @@ static int gd_handle_touch_debug(const ST__REMOTECONTROL_PACKET *packet, int *tx
     return tx_index;
 }
 
-static int gd_handle_no_backtel(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option)
+static int gd_handle_no_backtel(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option)
 {
     int noBacktel_mode;
 
@@ -124,7 +124,7 @@ static int gd_handle_no_backtel(const ST__REMOTECONTROL_PACKET *packet, int *tx_
     return tx_index;
 }
 
-static int gd_handle_map_init(const ST__REMOTECONTROL_PACKET *packet, int *tx_buf, int tx_index, int option)
+static int gd_handle_map_init(const ST__REMOTECONTROL_PACKET *packet, uint8_t *tx_buf, int tx_index, int option)
 {
     int RL;  // 옵션 이후 데이터
