+++ src/2__cm3/source/ble/tdc_ble_general_debug.h
@@ -19,6 +19,6 @@
  *  - 반환값   : 응답을 채운 뒤의 tx_index
  */
 int tdc_ble_general_debug_handle(const ST__REMOTECONTROL_PACKET *packet,
-                                    int *tx_buf, int tx_index);
+                                    uint8_t *tx_buf, int tx_index);
 
 #endif /* __tdc_ble_general_debug_h__ */
